refactor(netcdf): extract plotting logic into NetCDFPlot (pyramids.netcdf._plot)
`NetCDF.plot` is now a thin facade that constructs a `NetCDFPlot` and calls
`.run()`, mirroring how `Dataset.plot` delegates to its `Analysis` engine.
About 1430 lines of variable-resolution, selector-resolution, faceting,
animation, lazy-read and curvilinear-coord logic move out of `netcdf.py` into
the new `pyramids/netcdf/_plot.py`.

Moved helpers (formerly private methods on `NetCDF`, now on `NetCDFPlot`):
plot body -> run; _validate_facet_dims; _build_facet_stack; _resolve_animate_dim;
_render_animate; _maybe_log_lazy_hint; _resolve_curvilinear_coords;
_resolve_time_dim_name; _resolve_level_dim_name; _resolve_member_dim_name;
_coerce_coord_spec; _squeeze_leading_axes; _coord_shapes_match;
_cf_coordinates_pair; _looks_like_x_then_y; _remove_colorbar. The
_CURVILINEAR_NAME_PAIRS and _LAZY_HINT_THRESHOLD_BYTES constants move with them.

`NetCDFPlot` is not re-exported from `pyramids.netcdf` — it is an implementation
detail. `NetCDF.plot` keeps its full public signature and docstring.
`_check_not_container` stays on `NetCDF` (also used by read_array).

Tests that referenced the moved private symbols are updated for the new import
paths only (patch targets, caplog logger name, NetCDFPlot._remove_colorbar);
a new tests/netcdf/test_plot_module.py adds delegation smoke tests. Pure move,
no behaviour change.
